refactor: rename create-route-context to route-context and update related imports
